fixing counters for remaining requests

+++ point.cpp
@@ -3,7 +3,7 @@
 
 Point::Point() {}
 
-Point::Point(int x, int y) {
+Point::Point(double x, double y) {
     this->x = x;
     this->y = y;
 }
